feat: implement fractional numbers
