+++ fluido-types/src/lib.rs
@@ -2,3 +2,4 @@ pub mod concentration;
 pub mod error;
 pub mod expr;
 pub mod fluid;
+pub mod number;
